refactor: restructure repository files under `quests` module and update imports for improved organization
