+++ src/dependencies/services.py
@@ -1,13 +1,20 @@
 from fastapi import Depends
 
-from src.dependencies.repositories import get_guild_repo, get_objective_repo, get_project_repo, get_quest_repo, get_user_repo, \
-    get_pin_repo, \
-    get_world_repo, get_reward_repo
+from src.dependencies.repositories import (
+    get_guild_repo,
+    get_objective_repo,
+    get_project_repo,
+    get_quest_repo,
+    get_user_repo,
+    get_pin_repo,
+    get_world_repo,
+    get_reward_repo
+)
 from src.repositories.guild import GuildRepository
-from src.repositories.objective import ObjectiveRepository
+from src.repositories.quests.objective import ObjectiveRepository
 from src.repositories.project import ProjectRepository
-from src.repositories.quest import QuestRepository
-from src.repositories.reward import RewardRepository
+from src.repositories.quests.quest import QuestRepository
+from src.repositories.quests.reward import RewardRepository
 from src.repositories.user import UserRepository
 from src.repositories.pin import PinRepository
 from src.repositories.world import WorldRepository
