+++ src/repositories/quests/objective.py
